requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
psycopg2-binary==2.9.5
alembic==1.10.3
python-dotenv==1.0.0
pydantic==1.10.7
scipy==1.10.1
numpy==1.24.2
python-multipart==0.0.6
gunicorn==20.1.0